move method to avoid fix intra-package dependency
wmg_processing tests relied on wmg.api package, which required otherwise
unnecessary pip packages (web-related)
