+++ backend/wmg/data/utils.py
@@ -1,10 +1,11 @@
 import logging
 import time
-from typing import List
+from typing import List, Dict
 
 import tiledb
 
 from backend.wmg.data.schemas.corpus_schema import OBS_ARRAY_NAME
+from backend.wmg.data.snapshot import WmgSnapshot
 
 
 def log_func_runtime(func):
@@ -46,3 +47,74 @@ def create_empty_cube(uri: str, schema):
     Create an empty cube with expected schema (dimensions and attributes) at given uri
     """
     tiledb.Array.create(uri, schema, overwrite=True)
+
+
+def find_dim_option_values(criteria: Dict, snapshot: WmgSnapshot, dimension: str) -> list:
+    """Find values for the specified dimension that satisfy the given filtering criteria,
+    ignoring any criteria specified for the given dimension."""
+
+    filter_options_criteria = dict(criteria)
+    # Remove gene_ontology_term_ids from the criteria as it is not an eligible cross-filter dimension.
+    filter_options_criteria.pop("gene_ontology_term_ids", None)
+
+    # depluralize `dimension` if necessary
+    dimension = dimension[:-1] if dimension[-1] == "s" else dimension
+
+    # each element  in `linked_filter_sets` corresponds to the set of filters linked to the attributes specified for a corresponding criteria key
+    linked_filter_sets = []
+
+    # `all_criteria_attributes` is the set of all attributes specified across all criteria
+    all_criteria_attributes = set()
+
+    for key in filter_options_criteria:
+        attrs = filter_options_criteria[key]
+
+        # depluralize `key` if necessary
+        key = key[:-1] if key[-1] == "s" else key
+
+        # ignore the criteria for the specified dimension
+        if key != dimension:
+            if isinstance(attrs, list):
+                if len(attrs) > 0:
+                    # prepend the key to each attribute value
+                    prefixed_attributes = [key + "__" + val for val in attrs]
+                    all_criteria_attributes = all_criteria_attributes.union(prefixed_attributes)
+
+                    # for each attribute (attr) in `prefixed_attributes`,
+                    # get the set of filters for the specified dimension that are linked to `attr`
+                    linked_filter_set = set()
+                    for attr in prefixed_attributes:
+                        if dimension in snapshot.filter_relationships[attr]:
+                            linked_filter_set = linked_filter_set.union(
+                                set(snapshot.filter_relationships[attr][dimension])
+                            )
+
+                    linked_filter_sets.append(linked_filter_set)
+            else:
+                if attrs != "":
+                    prefixed_attribute = key + "__" + attrs
+                    all_criteria_attributes.add(prefixed_attribute)
+                    if dimension in snapshot.filter_relationships[prefixed_attribute]:
+                        linked_filter_sets.append(set(snapshot.filter_relationships[prefixed_attribute][dimension]))
+
+    # the candidate options are the intersection of the sets of linked filters for each criteria key
+    if len(linked_filter_sets) > 1:
+        candidate_options = linked_filter_sets[0].intersection(*linked_filter_sets[1:])
+    else:
+        candidate_options = linked_filter_sets[0]
+
+    # each valid option MUST be linked to at least one attribute specified in the criteria
+    # otherwise, there will be no data to display if that particular option is selected because
+    # the intersection will be null.
+    valid_options = []
+    for v in candidate_options:
+        loop_back_options = snapshot.filter_relationships[v]
+        all_loop_back_options = []
+        for dim in loop_back_options:
+            all_loop_back_options.extend(loop_back_options[dim])
+
+        if len(set(all_loop_back_options).intersection(all_criteria_attributes)) > 0:
+            valid_options.append(v)
+
+    # remove the prefix from each valid option and return the result
+    return [i.split("__")[1] for i in valid_options]
